refactor($piano-roll): rename PianoRollSpectrogram
- into NoteDisplaySpectrogram
