+++ src/renderer/components/piano-roll-zone/NoteDisplaySpectrogram.vue
